requirements.txt
python-telegram-bot==20.6
flask
flask-cors
python-dotenv
requests
gunicorn
watchdog
pymongo
requests